--- src/main/java/com/kichen/creation/commerce/exception/product/NotEnoughStockException.java
@@ -1,7 +1,0 @@
-package com.kichen.creation.commerce.exception.product;
-
-public class NotEnoughStockException extends RuntimeException {
-    public NotEnoughStockException(String message) {
-        super(message);
-    }
-}
