--- src/main/java/com/kichen/creation/commerce/exception/product/ProductNotFoundException.java
@@ -1,11 +1,0 @@
-package com.kichen.creation.commerce.exception.product;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-
-    public ProductNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
