+++ src/main/java/com/kichen/creation/commerce/product/exception/NotEnoughStockException.java
@@ -1,0 +1,7 @@
+package com.kichen.creation.commerce.product.exception;
+
+public class NotEnoughStockException extends RuntimeException {
+    public NotEnoughStockException(String message) {
+        super(message);
+    }
+}
