+++ src/main/java/com/kichen/creation/commerce/product/exception/ProductNotFoundException.java
@@ -1,0 +1,11 @@
+package com.kichen.creation.commerce.product.exception;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(String message) {
+        super(message);
+    }
+
+    public ProductNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
