+++ src/main/java/com/kichen/creation/commerce/product/repository/ProductRepository.java
@@ -1,0 +1,6 @@
+package com.kichen.creation.commerce.product.repository;
+
+import com.kichen.creation.commerce.product.domain.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {}
