--- src/main/java/com/kichen/creation/commerce/response/SuccessResponse.java
@@ -1,11 +1,0 @@
-package com.kichen.creation.commerce.response;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class SuccessResponse<T> {
-    private final int statusCode;
-    private final T data;
-}
